! some things scruitinizer found

+++ sources/ElkArte/Controller/Avatars.php
@@ -380,7 +380,7 @@ class Avatars
 			$extension = 'png';
 			$sizes[2] = IMAGETYPE_PNG;
 		}
-		$preferred_format = array_search($extension, $valid_avatar_extensions);
+		$preferred_format = (int) array_search($extension, $valid_avatar_extensions);
 		$mime_type = getValidMimeImageType($sizes[2]);
 
 		// Generate the final name
